# internal/api/generated/api.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CreateRecordRequest defines model for CreateRecordRequest.
type CreateRecordRequest struct {
	Artist      string  `json:"artist"`
	Condition   *string `json:"condition,omitempty"`
	Country     *string `json:"country,omitempty"`
	Format      *string `json:"format,omitempty"`
	Label       *string `json:"label,omitempty"`
	Notes       *string `json:"notes,omitempty"`
	ReleaseYear *int    `json:"releaseYear,omitempty"`
	Title       string  `json:"title"`
}

// Record defines model for Record.
type Record struct {
	Artist      string             `json:"artist"`
	Condition   *string            `json:"condition,omitempty"`
	Country     *string            `json:"country,omitempty"`
	CreatedAt   time.Time          `json:"createdAt"`
	Format      *string            `json:"format,omitempty"`
	Id          openapi_types.UUID `json:"id"`
	Label       *string            `json:"label,omitempty"`
	Notes       *string            `json:"notes,omitempty"`
	ReleaseYear *int               `json:"releaseYear,omitempty"`
	Title       string             `json:"title"`
	UpdatedAt   time.Time          `json:"updatedAt"`
}

// UpdateRecordRequest defines model for UpdateRecordRequest.
type UpdateRecordRequest = CreateRecordRequest

// RecordId defines model for RecordId.
type RecordId = openapi_types.UUID

// GetRecordsParams defines parameters for GetRecords.
type GetRecordsParams struct {
	// Q Search term for title, artist, label, or notes
	Q *string `form:"q,omitempty" json:"q,omitempty"`
}

// PostRecordsJSONRequestBody defines body for PostRecords for application/json ContentType.
type PostRecordsJSONRequestBody = CreateRecordRequest

// PutRecordsIdJSONRequestBody defines body for PutRecordsId for application/json ContentType.
type PutRecordsIdJSONRequestBody = UpdateRecordRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List or search records
	// (GET /records)
	GetRecords(c *gin.Context, params GetRecordsParams)
	// Add a record
	// (POST /records)
	PostRecords(c *gin.Context)
	// Delete a record
	// (DELETE /records/{id})
	DeleteRecordsId(c *gin.Context, id RecordId)
	// Describe a record
	// (GET /records/{id})
	GetRecordsId(c *gin.Context, id RecordId)
	// Update a record
	// (PUT /records/{id})
	PutRecordsId(c *gin.Context, id RecordId)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetRecords operation middleware
func (siw *ServerInterfaceWrapper) GetRecords(c *gin.Context) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetRecordsParams

	// ------------- Optional query parameter "q" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "q", c.Request.URL.Query(), &params.Q, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter q: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetRecords(c, params)
}

// PostRecords operation middleware
func (siw *ServerInterfaceWrapper) PostRecords(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostRecords(c)
}

// DeleteRecordsId operation middleware
func (siw *ServerInterfaceWrapper) DeleteRecordsId(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id RecordId

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteRecordsId(c, id)
}

// GetRecordsId operation middleware
func (siw *ServerInterfaceWrapper) GetRecordsId(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id RecordId

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetRecordsId(c, id)
}

// PutRecordsId operation middleware
func (siw *ServerInterfaceWrapper) PutRecordsId(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id RecordId

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PutRecordsId(c, id)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/records", wrapper.GetRecords)
	router.POST(options.BaseURL+"/records", wrapper.PostRecords)
	router.DELETE(options.BaseURL+"/records/:id", wrapper.DeleteRecordsId)
	router.GET(options.BaseURL+"/records/:id", wrapper.GetRecordsId)
	router.PUT(options.BaseURL+"/records/:id", wrapper.PutRecordsId)
}